+++ django_project/geosight/reference_dataset/api/v1/entities.py
@@ -23,6 +23,7 @@ from rest_framework import viewsets
 from rest_framework.renderers import JSONRenderer, BrowsableAPIRenderer
 
 from core.api_utils import common_api_params, ApiParams
+from core.pagination import Pagination, GeojsonPagination
 from core.renderers import GeoJSONRenderer
 from geosight.data.api.v1.base import BaseApiV1
 from geosight.georepo.models.entity import Entity
@@ -116,6 +117,14 @@ class EntityReferenceDatasetViewSet(BaseApiV1, viewsets.ReadOnlyModelViewSet):
     lookup_field = 'geom_id'
     lookup_value_regex = '[^/]+'
     renderer_classes = [GeoJSONRenderer, JSONRenderer, BrowsableAPIRenderer]
+
+    @property
+    def pagination_class(self):
+        """Return dynamic pagination class based on format."""
+        format = self.request.query_params.get('format')
+        if format == 'geojson':
+            return GeojsonPagination
+        return Pagination
 
     def get_serializer_class(self):
         """
